\name{GenHTSAHtmlRowUnit}
\alias{GenHTSAHtmlRowUnit}
\title{
Generate html codes for one unit of a row of a table
}
\description{
This is an internal function.
}
\usage{
GenHTSAHtmlRowUnit(content, href.attr=NULL,td.attr=NULL,header=FALSE)
}
\arguments{
\item{content}{
}
\item{href.attr}{
}
\item{td.attr}{
}
\item{header}{
}
}
\details{
This is an internal function.
}
\author{
Xin Wang
}

\keyword{internal}
