\name{paraCheck}
\alias{paraCheck}
\title{
Check parameters
}
\description{
This is an internal function.
}
\usage{
paraCheck(name, para)
}
\arguments{
\item{name}{
a single character value specifying the name of the variable to check
}
\item{para}{
the argument to check
}
}
\details{
This is an internal function.
}
\author{
Xin Wang
}

\keyword{internal}
