\name{permutationPvalueGsea}
\alias{permutationPvalueGsea}
\title{
Compute the nominal p-value associated with a GSEA
}
\description{
Compute the nominal p-value associated with a GSEA for one gene set
}
\usage{
permutationPvalueGsea(permutationsMatrix, geneList, geneSet, exponent, observedES)
}
\arguments{
  \item{permutationsMatrix}{
A matrix of permutations generated by "permutationsGenerator"
}
  \item{geneList}{
A geneList as in gseaScores (named vector, ordered, no duplicates nor NAs)
}
  \item{geneSet}{
A geneSet as in gseaScores
}
  \item{exponent}{
An exponent passed to gseaScores
}
  \item{observedES}{
The score obtained from gseaScores with the real data
}
}
\value{
A single p value
}
\references{
Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L., Gillette, M. A., Paulovich, A., Pomeroy, S. L., Golub, T. R., Lander, E. S. & Mesirov, J. P. (2005)  \emph{Gene set enrichment analysis: A knowledge-based approach for interpreting genome-wide expression profiles.} Proc. Natl. Acad. Sci. USA 102, 15545-15550.
}
\author{
Camille Terfve
}
\seealso{
gseaScores
}
\examples{
gl<-runif(100,min=0,max=5)
gl<-order(gl)
names(gl)<-as.character(sample(x=seq(from=1,to=100,by=1),size=100,replace=FALSE))
gs1<-sample(names(gl),size=20,replace=FALSE)
glgs1GSEA<-gseaScores(geneList=gl,geneSet=gs1,exponent=1,mode="score")
perm<-permutationsGeneratorGsea(nPermutations=1000,geneList=gl)
glgs1GSEA.pval<-permutationPvalueGsea(permutationsMatrix=perm,geneList=gl,
geneSet=gs1, exponent=1, observedES=glgs1GSEA)
}
