\name{writeHTSAHtmlHead}
\alias{writeHTSAHtmlHead}
\title{
Write the head part of report htmls
}
\description{
This is an internal function.
}
\usage{
writeHTSAHtmlHead(experimentName, htmlfile, rootdir=".")
}
\arguments{
\item{experimentName}{
a single character value specifying the name of the experiment (just
for you own record)
}
\item{htmlfile}{
a single character value specifying the name of the html file where
the head message will be stored
}
\item{rootdir}{
a single character value specifying the relative path from current 
html file to the root directory
}
}
\details{
This is an internal function.
}
\author{
Xin Wang
}

\keyword{internal}
