\name{writeHTSAHtmlSummary}
\alias{writeHTSAHtmlSummary}
\title{
Write the summary part of report htmls
}
\description{
This is an internal function.
}
\usage{
writeHTSAHtmlSummary(gsca=NULL, nwa=NULL, htmlfile)
}
\arguments{
\item{gsca}{
}
\item{nwa}{
}
\item{htmlfile}{
}
}
\details{
This is an internal function.
}
\author{
Xin Wang
}

\keyword{internal}
