\name{writeHTSAHtmlTab}
\alias{writeHTSAHtmlTab}
\title{
Write tabs to report htmls
}
\description{
This is an internal function.
}
\usage{
writeHTSAHtmlTab(enrichmentAnalysis,tab=c("GSCA","NWA"),htmlfile,rootdir=".",index=TRUE)
}
\arguments{
\item{enrichmentAnalysis}{
}
\item{tab}{
}
\item{htmlfile}{
}
\item{rootdir}{
}
\item{index}{
}
}
\details{
This is an internal function.
}
\author{
Xin Wang
}

\keyword{internal}
