\name{writeHTSAHtmlTable}
\alias{writeHTSAHtmlTable}
\title{
Write a table to report htmls
}
\description{
This is an internal function.
}
\usage{
writeHTSAHtmlTable(dat.tab, href.tab=NULL, signif.tab=NULL, row.attr.tab=NULL, tab.class,tab.name, htmlfile)
}
\arguments{
\item{dat.tab}{
}
\item{href.tab}{
}
\item{signif.tab}{
}
\item{row.attr.tab}{
}
\item{tab.class}{
}
\item{tab.name}{
}
\item{htmlfile}{
}
}
\details{
This is an internal function.
}
\author{
Xin Wang
}

\keyword{internal}
