\name{writeHTSAHtmlTail}
\alias{writeHTSAHtmlTail}
\title{
Write the tail part of report htmls
}
\description{
This is an internal function.
}
\usage{
writeHTSAHtmlTail(htmlfile)
}
\arguments{
\item{htmlfile}{
a single character value specifying the name of the html file where
the head message will be stored
}
}
\details{
This is an internal function.
}
\author{
Xin Wang
}

\keyword{internal}
