\name{GenHTSAHtmlRowUnit}
\alias{GenHTSAHtmlRowUnit}
\title{
Generate html codes for one unit of a row of a table
}
\description{
This is an internal function.
}
\usage{
GenHTSAHtmlRowUnit(content, href.attr=NULL,td.attr=NULL,header=FALSE)
}
\arguments{
\item{content}{
the content of the table unit
}
\item{href.attr}{
attributes of hyperlinks
}
\item{td.attr}{
attributes of a standard cell
}
\item{header}{
a logical value specifying whether the unit is a table header unit
}
}
\details{
This is an internal function.
}
\author{
Xin Wang
}

\keyword{internal}
