\name{htmlAttrVectorPaste}
\alias{htmlAttrVectorPaste}
\title{
Collapse an attribute vector for a table unit
}
\description{
This is an internal function.
}
\usage{
htmlAttrVectorPaste(vec)
}
\arguments{
\item{vec}{
}
}
\details{
This is an internal function.
}
\author{
Xin Wang
}

\keyword{internal}
