\name{makeGSEAplots}
\alias{makeGSEAplots}
\title{
Produce GSEA plots
}
\description{
This is an internal function.
}
\usage{
makeGSEAplots(geneList, geneSet, exponent, filepath, filename, 
output='png', ...)
}
\arguments{
\item{geneList}{
}
\item{geneSet}{
}
\item{geneSet}{
}
\item{exponent}{
}
\item{filepath}{
}
\item{filename}{
}
\item{output}{
}
\item{...}{
}
}
\details{
This is an internal function.
}
\author{
Benilton Carvalho, Xin Wang
}

\keyword{internal}
