\name{makeOverlapTable}
\alias{makeOverlapTable}
\title{
Find and save the overlapped genes between a gene set and hit list
}
\description{
This is an internal function.
}
\usage{
makeOverlapTable(geneSet, hits, mapID, filepath, filename)
}
\arguments{
\item{geneSet}{
}
\item{hits}{
}
\item{mapID}{
}
\item{filepath}{
}
\item{filename}{
}
}
\details{
This is an internal function.
}
\author{
Benilton Carvalho, Xin Wang
}

\keyword{internal}
