\name{writeHTSAHtmlHead}
\alias{writeHTSAHtmlHead}
\title{
Write the head part of report htmls
}
\description{
This is an internal function.
}
\usage{
writeHTSAHtmlHead(experimentName, htmlfile, rootdir=".")
}
\arguments{
\item{experimentName}{
}
\item{htmlfile}{
}
\item{rootdir}{
}
}
\details{
This is an internal function.
}
\author{
Xin Wang
}

\keyword{internal}
