\name{writeHTSAHtmlSummary}
\alias{writeHTSAHtmlSummary}
\title{
Write the summary part of report htmls
}
\description{
This is an internal function.
}
\usage{
writeHTSAHtmlSummary(gsca=NULL, nwa=NULL, htmlfile)
}
\arguments{
\item{gsca}{
an object of class \code{GSCA}
}
\item{nwa}{
an object of class \code{NWA}
}
\item{htmlfile}{
a single character value specifying the name of the html file where
the head message will be stored
}
}
\details{
This is an internal function.
}
\author{
Xin Wang
}

\keyword{internal}
