\name{writeHTSAHtmlTab}
\alias{writeHTSAHtmlTab}
\title{
Write tabs to report htmls
}
\description{
This is an internal function.
}
\usage{
writeHTSAHtmlTab(enrichmentAnalysis,tab=c("GSCA","NWA"),htmlfile,rootdir=".",index=TRUE)
}
\arguments{
\item{enrichmentAnalysis}{
}
\item{tab}{
a character vector specifying which tab to generate, "GSCA", "NWA" or both
}
\item{htmlfile}{
a single character value specifying the name of the html file where
the head message will be stored
}
\item{rootdir}{
a single character value specifying the relative path from current 
html file to the root directory
}
\item{index}{
a single logical value specifying whether to add an index tab or not
}
}
\details{
This is an internal function.
}
\author{
Xin Wang
}

\keyword{internal}
