\name{writeHTSAHtmlTable}
\alias{writeHTSAHtmlTable}
\title{
Write a table to report htmls
}
\description{
This is an internal function.
}
\usage{
writeHTSAHtmlTable(dat.tab, href.tab=NULL, signif.tab=NULL, row.attr.tab=NULL, tab.class,tab.name, htmlfile)
}
\arguments{
\item{dat.tab}{
a matrix storing the data of the table
}
\item{href.tab}{
a 3D array storing the hyperlinks of the data table
}
\item{signif.tab}{
a matrix specifying whether or not the data table will be highlighted 
}
\item{row.attr.tab}{
a matrix storing the html attributes of the table
}
\item{tab.class}{
a single character value specifying the class of the table
}
\item{tab.name}{
a single character value specifying the name of the table
}
\item{htmlfile}{
a single character value specifying the name of the html file where
the head message will be stored
}
}
\details{
This is an internal function.
}
\author{
Xin Wang
}

\keyword{internal}
